fix: make it works on worker
