+++ src/index.ts
@@ -1,10 +1,11 @@
 import { enableReactOptimization } from './config';
 import { createLink } from './createLink';
+// import { discoverProjectStyles } from './discoverProjectStyles';
 import { getCriticalRules, extractCriticalRules, getCriticalStyles, getUsedStyles } from './getCSS';
+import { loadStyleDefinitions, parseProjectStyles } from './loadStyleDefinitions';
 import { alterProjectStyles } from './operations';
 import { createCriticalStyleStream } from './reporters/critical';
 import { createStyleStream } from './reporters/used';
-import { discoverProjectStyles, loadStyleDefinitions, parseProjectStyles } from './scanForStyles';
 
 import { createUsedFilter as createUsedSelectorsFilter } from './utils/cache';
 
@@ -13,7 +14,7 @@ export { UsedTypes, StyleDefinition, SelectionFilter } from './types';
 export {
   createUsedSelectorsFilter,
   loadStyleDefinitions,
-  discoverProjectStyles,
+  // discoverProjectStyles,
   parseProjectStyles,
   alterProjectStyles,
   getUsedStyles,
